Create Initialize-Snort.sh
gets username and stores it as a variable for the path, updates, downloads my master repo, installs snort, verifies version to tell whether its going to need to use a lua file or a conf file. uses my rules from the snortrulescheatsheet.txt and appends the to the local.rules file. having trouble updating the /etc/snort/snort.conf or snort.lua HOME_NET variable. working on it