Create LocalHost-Containment.ps1
Locally run script using power shell for a malware infected host. To prevent ransomware or a worm from propagating the network for example. Script has the ability to disable all network adapters, preventing communication. It also prompts the admin if they would create a firewall rule to deny all outgoing connections with the exception of the current remote desktop protocol session. Once this is contained, you have the opportunity to find the malware variant and explore eradication.