Create Get-KeyManagement.ps1
OpenSSL is a pain sometimes. so i worked on this AI assisted creation of a key creation and encryption script. Generates a symmetric key in base64 or not. then asks if you want to encrypt the symmetric key with a recipients public key (public key needs to be in xml format) then does so accordingly. lastly, asks if you would like to generate your own rsa 2048 key pair.